% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letters2numbers.R
\name{letters2numbers}
\alias{letters2numbers}
\title{letters2numbers}
\usage{
letters2numbers(letterlist)
}
\arguments{
\item{letterlist}{A letter or vector of letters to get column numbers for.}
}
\value{
numeric vector useful for getting variable names via indexing
}
\description{
letters2numbers returns the number(s) corresponding to a letter or set of letters. Useful when a researcher tells you he is interested in column k, bc, and dd from an excel sheet.
}
\examples{
#indexLocation <- letters2numbers(c('k', 'bc', 'dd'))
#actualNames <- names(df)[indexLocation]
}

