% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qprint.R
\name{qprint}
\alias{qprint}
\title{qprint}
\usage{
qprint(table = FALSE, filename = "RPrintout.doc",
  tableHeading = "Table X", row.names = FALSE, fontsize = 10,
  column.margins = 1)
}
\arguments{
\item{table}{Either a data.frame, matrix, or table R object}

\item{filename}{Name of the output word document}

\item{tableHeading}{Heading for the table in the output word document}

\item{row.names}{If TRUE, the row.names of the input object is retained for the output document}

\item{fontsize}{Fontsize for the output document}

\item{column.margins}{Width of the columns in the output table}
}
\description{
qprint prints data.frame, matrix, and table objects to word documents.
}
\examples{
#qprint(table = OutputTable, filename = "OutputFileName.doc")
}

