% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_.R
\name{rep_}
\alias{rep_}
\title{rep_}
\usage{
rep_(x, times = 1, each = NULL)
}
\arguments{
\item{x}{Vector}

\item{times}{Number of times to repeat the vector}

\item{each}{Number of times to repeat each element in the vector}
}
\value{
A vector
}
\description{
rep_  is a 'better' implementation of the rep (repeat) base R function
}
\examples{
#rep_(c("A","B","C"),each=c(1,2,3))
#lst<-c(letters[1:4])
#rep(lst,times=c(1,1,4,4,3,3,2,2),each=2)
#rep_(lst,times=2,each=c(1,4,3,2))
#x<-letters[1:4]
#rep(x,times=3)
#rep_(x,times=3)
#rep(x,each=3)
#rep_(x,each=3)
#rep(x, each=c(1,2,3,4))
#rep(x, times=c(1,2,3,4))
#rep_(x, each=c(1,2,3,4))
#rep(x,times=c(1,2,3,4),each=2)
#rep_(x,times=2,each=c(1,2,3,4))
}

