% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdate.R
\name{tdate}
\alias{tdate}
\title{tdate}
\usage{
tdate(format = "\%Y\%B\%d")
}
\arguments{
\item{format}{Character string providing formatting instructions to format()}
}
\value{
character string of the current day's date
}
\description{
tdate returns a character string of the current day's date. It is a convenience wrapper for `format(Sys.Date())`.
}
\examples{
#tdate()
#tdate("\%Y\%m\%d")
#tdate("\%A, \%B \%d, \%Y")
}

